re-organized packages and added separate objects for commands
