+++ src/main/scala/com/tuplejump/snackfs/cassandra/model/GenericOpSuccess.scala
@@ -16,9 +16,6 @@
  * limitations under the License.
  *
  */
-package com.tuplejump.snackfs.model
+package com.tuplejump.snackfs.cassandra.model
 
-import org.apache.cassandra.thrift.Cassandra.AsyncClient
-import org.apache.thrift.transport.TNonblockingSocket
-
-case class ThriftClientAndSocket(client: AsyncClient, socket: TNonblockingSocket)
+case class GenericOpSuccess(){}
