+++ src/test/scala/com/tuplejump/snackfs/fs/model/INodeSpec.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  *
  */
-package com.tuplejump.snackfs.model
+package com.tuplejump.snackfs.fs.model
 
 import java.util.UUID
 import org.apache.hadoop.fs.Path
